requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
environ==1.0
python-environ==0.4.54
sqlparse==0.5.1
tzdata==2024.2
